+++ packages/boilerplate/composables/src/useFacet/index.ts
@@ -1,8 +1,8 @@
-import { Context, useFacetFactory } from '@vue-storefront/core';
+import { Context, useFacetFactory, FacetSearchResult } from '@vue-storefront/core';
 
 const factoryParams = {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  search: async (context: Context, params) => {
+  search: async (context: Context, params: FacetSearchResult<any>) => {
     console.log('Mocked: searchFacet');
     return {};
   }